+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* `createProxyHandler` (new deprecated) is replaced with `proxyHandler`.
+
+* Updated to be compatible with RFC 2616 Proxy specification.
+
 ## 0.0.2
 
 * Updated `README.md` and doc comments on `createProxyHandler`.